Use configured stdout & stderr objects for ssh
In https://github.com/PagerDuty/blender/pull/53, the output from
multi-ssh got separated by host in a hash. However, that hash never made
it back to the user because whatever the user passed in as stdout and
stderr got overwritten with a new hash reference. This will allow the
user to pass in an empty hash for stdout and stderr, and have their
command output populate the input hashes.